@@ -1,19 +1,21 @@
 source 'https://rubygems.org'
-ruby '2.4.1'
+ruby '2.4.2'
 
 git_source(:github) do |repo_name|
   repo_name = "#{repo_name}/#{repo_name}" unless repo_name.include?('/')
   "https://github.com/#{repo_name}.git"
 end
 
-gem 'acts_as_paranoid', '~> 0.5.0'
+gem 'acts_as_paranoid',
+    git: 'https://github.com/ActsAsParanoid/acts_as_paranoid.git',
+    ref: '085b5cc'
 gem 'after_party'
 gem 'bitters'
 gem 'bourbon'
 gem 'coffee-rails', '~> 4.2'
 gem 'decent_exposure'
 gem 'devise'
-gem 'factory_girl_rails'
+gem 'factory_bot_rails'
 gem 'faker'
 gem 'font-awesome-rails'
 gem 'jbuilder', '~> 2.5'
@@ -23,7 +25,7 @@ gem 'mail_interceptor'
 gem 'pg', '~> 0.18'
 gem 'pry'
 gem 'puma', '~> 3.0'
-gem 'rails', '~> 5.0.4'
+gem 'rails', '~> 5.1.4'
 gem 'record_tag_helper'
 gem 'rolify'
 gem 'sass-rails', '~> 5.0'
@@ -36,14 +38,14 @@ group :development, :test do
   gem 'byebug', platform: :mri
   gem 'capybara'
   gem 'dotenv-rails'
-  gem 'rspec-rails', '~> 3.6'
-  gem 'rubocop', '~> 0.49.1', require: false
+  gem 'rspec-rails', '~> 3.7'
+  gem 'rubocop', '~> 0.52', require: false
 end
 
 group :development do
   gem 'guard-rspec', require: false
   gem 'letter_opener'
-  gem 'listen', '~> 3.0.5'
+  gem 'listen'
   gem 'spring'
   gem 'spring-commands-rspec'
   gem 'spring-watcher-listen', '~> 2.0.0'
