@@ -1,12 +1,12 @@
 # users who have signed up, but done nothing yet
-FactoryGirl.create_list :user, 30
+FactoryBot.create_list :user, 30
 
 # a lot of user might sign up, log their first deed, and then forget about the
 # project
-FactoryGirl.create_list :user, 15, :with_one_deed
+FactoryBot.create_list :user, 15, :with_one_deed
 
 # these are the consistent users
-FactoryGirl.create_list :user, 5, :with_deeds
+FactoryBot.create_list :user, 5, :with_deeds
 
 # users with admin privelidges
-FactoryGirl.create_list :admin, 3
+FactoryBot.create_list :admin, 3
