@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   # TODO: implement this via Faker (will need to add a module for this to Faker)
   PRIZE_TITLES ||= [
     'Two free drinks at local pub',
