@@ -2,6 +2,6 @@ require 'rails_helper'
 
 RSpec.describe 'Linters' do
   it 'All factories are valid' do
-    FactoryGirl.lint
+    FactoryBot.lint
   end
 end
