@@ -28,7 +28,7 @@ RSpec::Matchers.define :show do |page_object_name|
     main_element = page.find('main')
     if main_element.present?
       classes = main_element[:class]
-      classes.present? ? classes : '(no classes applied to main element)'
+      classes.presence || '(no classes applied to main element)'
     else
       '(no main element found)'
     end
